Moves all components in DetailPage to its own file; Moves all typedefs in DetailPage to its own files
