+++ react/src/features/marketplace/detail/typedefs/CommentType.ts
@@ -1,0 +1,7 @@
+export type CommentType = {
+  profileImgUrl: string,
+  authorName: string,
+  date: string,
+  rating: number,
+  comment: string,
+};
