+++ react/src/features/marketplace/detail/typedefs/MarketplaceDetailCardType.ts
@@ -1,0 +1,6 @@
+export type MarketplaceDetailCardType = {
+  authorName: string,
+  publishedDate: string,
+  totalSaved: number,
+  rating: number,
+};
