+++ react/src/features/marketplace/detail/typedefs/MarketplaceDetailType.ts
@@ -1,0 +1,9 @@
+import { CommentType } from './CommentType';
+import { MarketplaceDetailCardType } from './MarketplaceDetailCardType';
+
+export type MarketplaceDetailType = {
+  name: string,
+  description: string,
+  detailCardData: MarketplaceDetailCardType,
+  commentsData: CommentType[]
+};
